More editor additions

+++ editor/songs.lua
@@ -19,6 +19,8 @@ local songOptions <const> = {
 local mapOptions <const> = {
     "New Map",
     "Edit Chart",
+    "Move Up",
+    "Move Down",
     "Rename",
     "Duplicate",
     "Delete"
@@ -26,7 +28,7 @@ local mapOptions <const> = {
 
 local guideText <const> = "Read the online level editor manual\nto get started making your own levels!"
 
-local drawerHeaders = {
+local drawerHeaders <const> = {
     songs = "Songs",
     maps = "Maps"
 }
@@ -34,6 +36,7 @@ local drawerHeaders = {
 -- define variables
 local function getListOfSongs()
     local songFiles = pd.file.listFiles("/custom songs/")
+    if songFiles == nil then songFiles = {} end
 
     for i=#songFiles,1,-1 do
         if songFiles[i]:sub(-4) ~= '.pda' then
@@ -47,7 +50,29 @@ local function getListOfSongs()
     return songFiles
 end
 
-local songList = getListOfSongs()
+songList = getListOfSongs()
+
+editorData = pd.datastore.read("editorData")
+if editorData == nil then editorData = {} end
+
+for i=1,#songList do
+    if editorData[songList[i]] == nil then
+        editorData[songList[i]] = {
+            songData = {
+                name = songList[i],
+                artist = "N/A",
+                bpm = 180,
+                bpmChanges = {},
+                beatOffset = 0,
+                preview = 0,
+                difficulties = {}
+            },
+            mapData = {}
+        }
+    end
+end
+
+pd.datastore.write(editorData, "editorData")
 
 qrCode = gfx.image.new("sprites/manualCode")
 
@@ -57,7 +82,7 @@ if editorData == nil then editorData = {} end
 local selecting = "song"
 
 local songSelection = 1
-local songSelectionRounded = songSelection
+songSelectionRounded = songSelection
 local oldSongSelection = songSelection
 
 local songOptionSelection = 1
@@ -67,26 +92,9 @@ local oldSongOptionSelection = songOptionSelection
 local deleting = false
 
 
-local function truncateString(string, width, font)
-    local stringWidth = gfx.getTextSize(string, font)
-    local newString = string
-
-    while stringWidth > width do
-        if newString:sub(-3) == "..." then
-            newString = newString:sub(1, newString:len()-4).."..."
-        else
-            newString = newString:sub(1, newString:len()-1).."..."
-        end
-        stringWidth = gfx.getTextSize(newString, font)
-    end
-
-    return newString
-end
-
-
 function updateEditorSongsSelect()
 
-    if not deleting then
+    if not deleting and #songList > 0 then
         if selecting == "song" then
             if bPressed or leftPressed then
                 sfx.low:play()
@@ -125,7 +133,7 @@ function updateEditorSongsSelect()
             end
             songOptionSelection += crankChange/90
         end
-    else
+    elseif deleting then
         if aHeld and upHeld then
             pd.file.delete("/custom songs/"..songList[songSelectionRounded]..".pda")
             songList = getListOfSongs()
@@ -135,6 +143,11 @@ function updateEditorSongsSelect()
         if bPressed or downPressed or leftPressed or rightPressed then
             deleting = false
             sfx.switch:play()
+        end
+    else
+        if bPressed or leftPressed then
+            sfx.low:play()
+            return "exitEditor"
         end
     end
     -- round song selection
@@ -226,40 +239,4 @@ function drawEditorSongSelect()
         gfx.drawTextAligned(warningText, screenCenterX, textY, kTextAlignment.center)
     end
 
-end
-
-function drawScrollingList(list, selectedItem, x, y, width, height, padding, font, fill, invert)
-    local textHeight = font[playdate.graphics.font.kVariantNormal]:getHeight()
-    local maxItems = math.floor(height/(textHeight+padding))
-    for i=math.max(selectedItem-maxItems/2+1, 1),#list do
-        local textY = y+(textHeight+padding)*(math.min(i, i-selectedItem+maxItems/2)-1)+padding
-        local textX = x+padding
-        
-        local truncatedText = truncateString(list[i], width, font)
-        local selectBoxColor = gfx.kColorBlack
-        local selectedTextColor = gfx.kDrawModeInverted
-        local textColor = gfx.kDrawModeCopy
-        if invert then
-            selectBoxColor = gfx.kColorWhite
-            selectedTextColor = gfx.kDrawModeCopy
-            textColor = gfx.kDrawModeInverted
-        end
-        if selectedItem == i then
-            local textWidth = gfx.getTextSize(truncatedText, font)
-            gfx.setColor(selectBoxColor)
-            if fill then
-                gfx.fillRoundRect(textX-padding, textY-padding, textWidth+padding*2, textHeight+padding*2, 3)
-                gfx.setImageDrawMode(selectedTextColor)
-                gfx.drawText(truncatedText, textX, textY, font)
-            else
-                gfx.drawRoundRect(textX-padding, textY-padding, textWidth+padding*2, textHeight+padding*2, 3)
-                gfx.setImageDrawMode(textColor)
-                gfx.drawText(truncatedText, textX, textY, font)
-            end
-        else
-            gfx.setImageDrawMode(textColor)
-            gfx.drawText(truncatedText, textX, textY, font)
-        end
-    end
-    gfx.setImageDrawMode(gfx.kDrawModeCopy)
 end
